fix: address latest PR feedback
